Fix saving of max attacks to display
The regex for detecting NaN was correct, but returned the wrong boolean value, it was backwards. This resulted in a valid entry being seen as invalid, and the default 100 was reverted to. Saved correctly now, and reflected in the dialog title. Although the dialog title does not dynamically update, yet.