Initial request queue implementation
In order to get around the 100 requests/min limitation, a primitive request queue is being implemented to (perhaps) force deferred request submission. This change queues requests in a variable length array, but thet are currently submitted in rapid succession as they always have been. A good algorithm for selective deferral still needs to be worked out.