Massive refactor
Moved a lot of code to sub-functions, moved global vars to top of file in one place, and start the script running via new function, 'go()' (equivalent to 'main()').

This is in preparation for adding some new functionality. Namely, 'clear all/select all' buttons for the configuration UI, and the ability to alphabetize the company lists, which will require changing the array of company names into a two-dimensional array, including company ID's associated with the names.