Prevent API calls while travelling
The base URL is the same for the main page (where we want to display) and the travel page. On the travel page, the clouds trigger the MutationObserver constantly, causing an API call; this will exceed our 100 per min limit, as well as being completely useless. So, we disconnect the observer and only re-connect if we are in Torn, or we wait until we've landed and check again.